chore: move box, distributions, mutant, and occlusions to new mutants submodule
