+++ rex_xai/multi_explanation.py
@@ -10,7 +10,7 @@ import torch as tt
 from itertools import combinations
 
 from rex_xai.explanation import Explanation
-from rex_xai.distributions import random_coords, Distribution
+from rex_xai.mutants.distributions import random_coords, Distribution
 from rex_xai.utils.logger import logger
 from rex_xai.utils._utils import powerset, clause_area, SpatialSearch
 from rex_xai.output.visualisation import save_multi_explanation, save_image, plot_image_grid
