+++ rex_xai/mutants/mutant.py
@@ -10,7 +10,7 @@ except ImportError:
     from anytree.search import find
 
 import matplotlib.pyplot as plt
-from rex_xai.box import Box
+from rex_xai.mutants.box import Box
 from rex_xai.prediction import Prediction
 from rex_xai.utils.logger import logger
 from rex_xai.input.input_data import Data
