+++ rex_xai/utils/_utils.py
@@ -9,7 +9,7 @@ import torch as tt
 import numpy as np
 from skimage.segmentation import mark_boundaries
 from rex_xai.utils.logger import logger
-from rex_xai.box import Box
+from rex_xai.mutants.box import Box
 
 Strategy = Enum("Strategy", ["Global", "Spatial", "MultiSpotlight", "Contrastive"])
