+++ tests/conftest.py
@@ -3,9 +3,9 @@ import torch as tt
 import pytest
 from cached_path import cached_path
 from rex_xai.utils._utils import get_device
-from rex_xai.box import initialise_tree
+from rex_xai.mutants.box import initialise_tree
 from rex_xai.input.config import CausalArgs, process_custom_script, Strategy
-from rex_xai.distributions import Distribution
+from rex_xai.mutants.distributions import Distribution
 from rex_xai.rex import (
     calculate_responsibility,
     get_prediction_func_from_args,
