+++ tests/unit_tests/config_test.py
@@ -3,7 +3,7 @@ import copy
 import pytest
 from rex_xai.utils._utils import Queue, Strategy
 from rex_xai.input.config import CausalArgs, process_config_dict, read_config_file
-from rex_xai.distributions import Distribution
+from rex_xai.mutants.distributions import Distribution
 
 
 @pytest.fixture
